requirements.txt
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
pandas==2.1.4
openpyxl==3.1.2
python-dotenv==1.0.0
colorama==0.4.6
tqdm==4.66.1
pyyaml==6.0.1
